# fix: add value indices on nft_custody views