@rafaelcr 2022-06-14
Notice it needs 2 workers for the "Before" query @CharlieC3