@blockstack-devops 2022-06-21
:tada: This PR is included in version 4.0.3 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@stacks/stacks-blockchain-api-types/v/4.0.3)
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@stacks/blockchain-api-client/v/4.0.3)
- [GitHub release](https://github.com/hirosystems/stacks-blockchain-api/releases/tag/v4.0.3)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: